# docs/swagger.yaml
basePath: /
definitions:
  traineeEVOFintech.Transaction:
    properties:
      amountOriginal:
        type: number
      amountTotal:
        type: number
      commissionClient:
        type: number
      commissionPS:
        type: number
      commissionProvider:
        type: number
      dateInput:
        type: string
      datePost:
        type: string
      partnerObjectId:
        type: integer
      payeeBankAccount:
        type: string
      payeeBankMfo:
        type: integer
      payeeId:
        type: integer
      payeeName:
        type: string
      paymentNarrative:
        type: string
      paymentNumber:
        type: string
      paymentType:
        type: string
      requestId:
        type: integer
      service:
        type: string
      serviceId:
        type: integer
      status:
        type: string
      terminalId:
        type: integer
      transactionId:
        type: integer
    type: object
host: localhost:8000
info:
  contact:
    email: rostislav.pylypiv@gmail.com
    name: Rostyslav Pylypiv
  description: API Server for Transactions Application
  title: Trainee EVO Fintech project (Transactions App API)
  version: "1.0"
paths:
  /get-csv:
    get:
      operationId: get_csv
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        type: string
      - description: Terminal ID
        in: query
        name: terminal_id
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Payment Type
        in: query
        name: payment_type
        type: string
      - description: 'Date Post From (Example: 2022-08-17)'
        in: query
        name: date_post_from
        type: string
      - description: 'Date Post To (Example: 2022-08-17)'
        in: query
        name: date_post_to
        type: string
      - description: 'Payment Narrative (Example: ''про надання послуг'')'
        in: query
        name: payment_narrative
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: getCSV
      tags:
      - Transactions
  /get-csv-file:
    get:
      operationId: get_csv_file
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        type: string
      - description: Terminal ID
        in: query
        name: terminal_id
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Payment Type
        in: query
        name: payment_type
        type: string
      - description: 'Date Post From (Example: 2022-08-17)'
        in: query
        name: date_post_from
        type: string
      - description: 'Date Post To (Example: 2022-08-17)'
        in: query
        name: date_post_to
        type: string
      - description: 'Payment Narrative (Example: ''про надання послуг'')'
        in: query
        name: payment_narrative
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: getCSVFile
      tags:
      - Transactions
  /get-json:
    get:
      operationId: get_json
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        type: string
      - description: Terminal ID
        in: query
        name: terminal_id
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Payment Type
        in: query
        name: payment_type
        type: string
      - description: 'Date Post From (Example: 2022-08-17)'
        in: query
        name: date_post_from
        type: string
      - description: 'Date Post To (Example: 2022-08-17)'
        in: query
        name: date_post_to
        type: string
      - description: 'Payment Narrative (Example: ''про надання послуг'')'
        in: query
        name: payment_narrative
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/traineeEVOFintech.Transaction'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: getJSON
      tags:
      - Transactions
  /upload-csv:
    post:
      consumes:
      - multipart/form-data
      operationId: upload_csv
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: uploadCSV
      tags:
      - Transactions
swagger: "2.0"